fix: exclude transitive jar dependencies provided by JPI plugins
Filter out jar files from the JPI classpath that are already provided by JPI plugin dependencies, even when those same jars are also needed as transitive dependencies of other non-JPI dependencies.

This ensures that when a JPI plugin provides a library (like Jackson) and another dependency also transitively depends on the same library, the library jars are not duplicated in the final JPI package.